% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TJU_Fiscal_Year.R
\name{TJU_Fiscal_Year}
\alias{TJU_Fiscal_Year}
\title{TJU Fiscal Year}
\usage{
TJU_Fiscal_Year(x)
}
\arguments{
\item{x}{\link[base]{integer} scalar}
}
\value{
Function \link{TJU_Fiscal_Year} returns a length-two \link[base]{Date} \link[base]{vector},
indicating the start (July 1 of the previous calendar year) and end date (June 30) of a fiscal year.
}
\description{
..
}
\examples{
TJU_Fiscal_Year(2022L)

}
