% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TJU_SchoolTerm.R
\name{TJU_SchoolTerm}
\alias{TJU_SchoolTerm}
\title{TJU School Term}
\usage{
TJU_SchoolTerm(x)
}
\arguments{
\item{x}{a \link[base]{Date} object}
}
\value{
Function \code{\link[=TJU_SchoolTerm]{TJU_SchoolTerm()}} returns a \link[base]{character} \link[base]{vector}
}
\description{
..
}
\examples{
c('2021-03-14', '2022-01-01', '2022-05-01') |>
 as.Date() |> 
 TJU_SchoolTerm()

}
