% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday.R
\name{workday}
\alias{workday}
\title{Workdays, Weekends and Holidays}
\usage{
workday(
  x,
  Holiday = c("USNewYearsDay", "USMLKingsBirthday", "USMemorialDay", "USIndependenceDay",
    "USLaborDay", "USThanksgivingDay", "USChristmasDay"),
  vacation
)
}
\arguments{
\item{x}{time-span to be summarized.  Acceptable objects are
\describe{
\item{\link[base]{character} scalar or \link[base]{vector} of format \code{'YYYY-MM'}}{e.g., \code{'2021-01'} for January 2021}
\item{\link[base]{character} scalar or \link[base]{vector} of format \code{'YYYY Q[1-4]'}}{e.g., \code{'2021 Q1'} for the first quarter of 2021 (January to March)}
\item{\link[base]{integer} scalar or \link[base]{vector}}{e.g., \code{2021L} for year 2021}
\item{\link[zoo]{yearqtr} and/or \link[zoo]{yearmon} objects}{e.g., \code{zoo::as.yearqtr('2025 Q2')} for the second quarter of 2025 (April to June)}
}}

\item{Holiday}{\link[base]{character} \link[base]{vector}, see function \link[timeDate]{holiday}}

\item{vacation}{\link[base]{Date} \link[base]{vector}, vacation days}
}
\value{
Function \code{\link[=workday]{workday()}} returns a \link[base]{factor}.
}
\description{
To summarize the number of workdays, weekends, holidays and vacations in a given time period,
e.g., a month or a quarter of a year.
}
\details{
Function \code{\link[=workday]{workday()}} summarizes the workdays, weekends,
Jefferson paid holidays
(New Year’s Day, Martin Luther King, Jr. Day, Memorial Day, Fourth of July, Labor Day, Thanksgiving and Christmas)
and your vacation (e.g., sick, personal, etc.) days (if any),
in a given time-span.

Per Jefferson policy (source needed),
if a holiday is on Saturday, then the preceding Friday is considered to be a weekend day.
If a holiday is on Sunday, then the following Monday is considered to be a weekend day.
}
\examples{
'2025 Q2' |> zoo::as.yearqtr() |> workday() |> table()
'2025 Q3' |> zoo::as.yearqtr() |> workday() |> table()
'2025 Q4' |> zoo::as.yearqtr() |> workday() |> table()

'2025 Q2' |> 
 zoo::as.yearqtr() |>
 workday(vacation = seq.Date(
  from = as.Date('2025-04-10'), to = as.Date('2025-04-24'), by = 1
 )) |>
 table()
 
2021L |> workday() |> table()

tryCatch(c('2019-10', '2019-12') |> zoo::as.yearmon() |> workday(), error = identity)
c('2019-10' |> zoo::as.yearmon() |> workday(), 
  '2019-12' |> zoo::as.yearmon() |> workday()) |> table() # work-around

}
\keyword{internal}
